+++ src/core/types.ts
@@ -69,8 +69,6 @@ export function convertToFrame(json: any) {
 }
 
 /**
- * Frame Request
- *
  * Note: exported as public Type
  */
 export type FrameButtonMetadata =
@@ -82,11 +80,10 @@ export type FrameButtonMetadata =
   | {
       action?: 'post' | 'post_redirect';
       label: string;
+      target?: string;
     };
 
 /**
- * Frame Request
- *
  * Note: exported as public Type
  */
 export type FrameInputMetadata = {
@@ -94,8 +91,6 @@ export type FrameInputMetadata = {
 };
 
 /**
- * Frame Request
- *
  * Note: exported as public Type
  */
 export type FrameImageMetadata = {
@@ -104,25 +99,35 @@ export type FrameImageMetadata = {
 };
 
 /**
- * Frame Request
- *
+ * Note: exported as public Type
+ */
+export type FrameMetadataReact = FrameMetadataType & {
+  ogDescription?: string;
+  ogTitle?: string;
+  wrapper?: React.ComponentType<any>;
+};
+
+/**
  * Note: exported as public Type
  */
 export type FrameMetadataType = {
+  // A list of strings which are the label for the buttons in the frame (max 4 buttons).
   buttons?: [FrameButtonMetadata, ...FrameButtonMetadata[]];
+  // An image which must be smaller than 10MB and should have an aspect ratio of 1.91:1
   image: string | FrameImageMetadata;
+  // The text input to use for the Frame.
   input?: FrameInputMetadata;
   /** @deprecated Prefer `postUrl` */
   post_url?: string;
+  // A valid POST URL to send the Signature Packet to.
   postUrl?: string;
   /** @deprecated Prefer `refreshPeriod` */
   refresh_period?: number;
+  // A period in seconds at which the app should expect the image to update.
   refreshPeriod?: number;
 };
 
 /**
- * Frame Metadata Response
- *
  * Note: exported as public Type
  */
 export type FrameMetadataResponse = Record<string, string>;
@@ -136,6 +141,7 @@ type EASAttesterAddress = Address;
 /**
  * Ethereum Attestation Service (EAS) Schema Uid
  * The schema identifier associated with the EAS attestation.
+ *
  * Note: exported as public Type
  */
 export type EASSchemaUid = Address;
@@ -143,6 +149,7 @@ export type EASSchemaUid = Address;
 /**
  * Ethereum Attestation Service (EAS) Attestation
  * GraphQL response for EAS Attestation
+ *
  * Note: exported as public Type
  */
 export type EASAttestation = {
@@ -160,6 +167,7 @@ export type EASAttestation = {
 /**
  * Ethereum Attestation Service (EAS) Chain Definition
  * The definition of a blockchain chain supported by EAS attestations.
+ *
  * Note: exported as public Type
  */
 export type EASChainDefinition = {
@@ -167,3 +175,26 @@ export type EASChainDefinition = {
   id: number; // blockchain source id
   schemaUids: EASSchemaUid[]; // Array of EAS Schema UIDs
 };
+
+/**
+ * Settings to simulate statuses on mock frames.
+ *
+ * Note: exported as public Type
+ */
+export type MockFrameRequestOptions = {
+  following?: boolean; // Indicates if the viewer clicking the frame follows the cast author
+  interactor?: {
+    fid?: number; // Viewer Farcaster ID
+    custody_address?: string; // Viewer custody address
+    verified_accounts?: string[]; // Viewer account addresses
+  };
+  liked?: boolean; // Indicates if the viewer clicking the frame liked the cast
+  recasted?: boolean; // Indicates if the viewer clicking the frame recasted the cast
+};
+
+/**
+ * A mock frame request payload
+ *
+ * Note: exported as public Type
+ */
+export type MockFrameRequest = FrameRequest & { mockFrameData: Required<FrameValidationData> };