+++ src/index.ts
@@ -4,6 +4,7 @@ export { getEASAttestations } from './core/getEASAttestations';
 export { getFrameHtmlResponse } from './core/getFrameHtmlResponse';
 export { getFrameMetadata } from './core/getFrameMetadata';
 export { getFrameMessage } from './core/getFrameMessage';
+export { getMockFrameRequest } from './core/getMockFrameRequest';
 export { FrameMetadata } from './components/FrameMetadata';
 export { Avatar } from './components/Avatar';
 export { Name } from './components/Name';
@@ -13,7 +14,10 @@ export type {
   FrameButtonMetadata,
   FrameImageMetadata,
   FrameInputMetadata,
+  FrameMetadataReact,
   FrameMetadataType,
   FrameRequest,
   FrameValidationData,
+  MockFrameRequest,
+  MockFrameRequestOptions,
 } from './core/types';