+++ src/xmtp/index.ts
@@ -1,0 +1,2 @@
+export { getXmtpFrameMessage } from './getXmtpFrameMessage';
+export { isXmtpFrameRequest } from './isXmtpFrameRequest';